requirements.txt
python-pptx==0.6.21
deep-translator==1.11.4
python-dotenv==1.0.1
tqdm==4.66.2
openai==1.65.2 